feat: extract hook
