+++ frontend/src/components/Carousel/Carousel.module.scss
@@ -13,7 +13,7 @@
 
 .pagination {
     position: absolute;
-    bottom: 30px;
+    bottom: 36px;
     left: 50%;
     transform: translateX(-50%);
 }
